refactor: carNameRender 함수 이동
