+++ js/controller/startGame.js
@@ -1,5 +1,4 @@
 import { store } from "../model.js";
-import { carNameRender } from "./carNameRender.js";
 import { renderWinner } from "./renderWinner.js";
 import {
   gameProcess,
@@ -7,6 +6,8 @@ import {
   deleteSpinnerIcon,
 } from "./gameProcess.js";
 import { winnerMessage } from "../view/winnerTemplate.js";
+import { $ } from "../constants/DOM.js";
+import { gameProcessTemplate } from "../view/gameProcessTemplate.js";
 
 export const startGame = () => {
   carNameRender();
@@ -22,4 +23,12 @@ export const startGame = () => {
       setTimeout(winnerMessage, 2000);
     }
   }, 1000);
+};
+
+export const carNameRender = () => {
+  const app = $("#app");
+  const div = document.createElement("div");
+
+  div.innerHTML = gameProcessTemplate(store.players);
+  app.appendChild(div);
 };
